Changed default http client timeout to 0
The expectation is that large files will be uploaded, so we don't want
those uploads to timeout prematurely.  And if the http client is hung
waiting for a response, it won't stop any of th other threads.

I may reconsider this, and set it to a really high value, like 6 hours
or something.